added enumeration of response data elements from weather forecast
simplified the logic to use all elements of the enumerated class,
added and index and api_name for the enum elements,

as enums are natively sorted, we don't need to additionally sort on
their index, but for sanity purpose, and failure tolerance we might want
to add that in later.

added centering & resizing, and `lift` to tk so that the application is
in the center of the screen, and on top of other applications when
launched, potential to add a "force focus" when starting to enhance UX.

When plots are added screen dynamically resizes to account for new
plots.